Fix code-check warnings related to Gtest functions.
- Suppress code-check warnings related to Gtest functions, for example:

```
Initialization of 'test_info_' with static storage duration may throw an exception that cannot be caught
```

- Fix some code format glitches, as reported by clang-format.

+++ libsrc/File.cc
@@ -68,7 +68,7 @@ std::string File::ModificationDate() {
 
   auto modTime = std::string(std::asctime(std::localtime(&timestamp)));
   // Clean any non-printable characters.
-  for (int i = 0; i < modTime.length(); i++) {
+  for (size_t i = 0; i < modTime.length(); i++) {
     if (!isprint(modTime[i])) {
       modTime[i] = ' ';
     }
